Extension; Creating the basic code to try and use it in behat
